chore(data-warehouse): relocate warehouse_sources commands and route wiring through facade
Moves the ws-owned management commands into the product (removing temporal-wiring couplings) and repoints the external temporal-wiring consumers (revenue/signals apps, stripe views, data-modeling, temporal worker bootstrap) at facade.temporal/facade.sources.
