+++ posthog/temporal/data_modeling/run_workflow.py
@@ -74,7 +74,7 @@ from products.data_warehouse.backend.s3 import ensure_bucket_exists, get_s3_clie
 from products.endpoints.backend.rate_limit import update_materialization_ready_for_saved_query
 from products.endpoints.backend.services.materialization import prepare_executable_query
 from products.warehouse_sources.backend.models.table import DataWarehouseTable
-from products.warehouse_sources.backend.temporal.data_imports.util import prepare_s3_files_for_querying
+from products.warehouse_sources.backend.facade.temporal import prepare_s3_files_for_querying
 
 LOGGER = get_logger(__name__)
