+++ products/data_modeling/backend/models/datawarehouse_saved_query.py
@@ -31,7 +31,7 @@ from products.warehouse_sources.backend.models.util import (
     clean_type,
     remove_named_tuples,
 )
-from products.warehouse_sources.backend.temporal.data_imports.naming_convention import NamingConvention
+from products.warehouse_sources.backend.facade.sources import NamingConvention
 
 logger = structlog.get_logger(__name__)
