+++ products/revenue_analytics/backend/views/sources/stripe/customer.py
@@ -15,7 +15,7 @@ from products.revenue_analytics.backend.views.sources.constants import (
 from products.revenue_analytics.backend.views.sources.helpers import extract_json_string, get_cohort_expr
 from products.warehouse_sources.backend.models.external_data_schema import ExternalDataSchema
 from products.warehouse_sources.backend.models.table import DataWarehouseTable
-from products.warehouse_sources.backend.temporal.data_imports.sources.stripe.constants import (
+from products.warehouse_sources.backend.facade.sources import (
     CHARGE_RESOURCE_NAME as STRIPE_CHARGE_RESOURCE_NAME,
     CUSTOMER_RESOURCE_NAME as STRIPE_CUSTOMER_RESOURCE_NAME,
     INVOICE_RESOURCE_NAME as STRIPE_INVOICE_RESOURCE_NAME,
