+++ products/revenue_analytics/backend/views/sources/stripe/revenue_item.py
@@ -14,7 +14,7 @@ from products.revenue_analytics.backend.views.sources.helpers import (
     is_zero_decimal_in_stripe,
 )
 from products.warehouse_sources.backend.models.table import DataWarehouseTable
-from products.warehouse_sources.backend.temporal.data_imports.sources.stripe.constants import (
+from products.warehouse_sources.backend.facade.sources import (
     CHARGE_RESOURCE_NAME as STRIPE_CHARGE_RESOURCE_NAME,
     INVOICE_RESOURCE_NAME as STRIPE_INVOICE_RESOURCE_NAME,
 )
