+++ products/warehouse_sources/backend/management/commands/generate_source_configs.py
