+++ products/warehouse_sources/backend/temporal/data_imports/sources/common/test/test_source_config_generator.py
@@ -17,7 +17,7 @@ from posthog.schema import (
     SourceFieldSwitchGroupConfig,
 )
 
-from posthog.management.commands.generate_source_configs import SourceConfigGenerator
+from products.warehouse_sources.backend.management.commands.generate_source_configs import SourceConfigGenerator
 
 from products.warehouse_sources.backend.temporal.data_imports.sources.common.base import FieldType
 from products.warehouse_sources.backend.types import ExternalDataSourceType
